+++ scripts/tests.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 if [ -x "$(which python)" ]; then
-    ./py2/bin/python -m unittest plumber.tests.test_plumber
+    ./py2/bin/python -m plumber.tests.__init__
 fi
 if [ -x "$(which python3)" ]; then
-    ./py3/bin/python -m unittest plumber.tests.test_plumber
+    ./py3/bin/python -m plumber.tests.__init__
 fi
 if [ -x "$(which pypy)" ]; then
-    ./pypy/bin/python -m unittest plumber.tests.test_plumber
+    ./pypy/bin/python -m plumber.tests.__init__
 fi